Add files via upload
Project on medical insurance charges using the below-mentioned parameters.
1. Age
Age of the insured. Integer quantity.
Gender
2. Gender of the insured. This parameter has been mapped to numerical values in the following way:
GenderAssigned ValueFemale1,Male2
3. BMI
Body mass index of the insured. Float value quantity.
4. No_of_Children
Number of children the insured person has. Integer quantity.
5. Smoker
Whether the insured person is a smoker or not. This parameter has been mapped to numerical values in the following way:
SmokerAssigned ValueSmoker1, Non-Smoker2
6. Region
Which region of the USA does the insured belong to. This parameter has been mapped to numerical values in the following way:
RegionAssigned ValueNorthwest1, Northeast2, Southwest 3, and Southeast 4
7. Charges
The charges for the insurance are in USD. Floating value quantity.
Link to dataset: https://lnkd.in/ds_WKaHu